+++ opendatalinter/csv_linter.py
@@ -11,24 +11,8 @@ import pandas as pd
 from jeraconv import jeraconv
 
 from .vo import LintResult, InvalidContent, InvalidCellFactory
+from .funcs import to_csv_format, before_check_1_1
 from .errors import TitleEstimateError, HeaderEstimateError
-
-
-def to_csv_format(txt):
-    output = StringIO()
-    writer = csv.writer(output)
-    writer.writerow(txt)
-    return output.getvalue()
-
-
-def before_check_1_1(func):
-    def wrapper(self, *args, **kwargs):
-        if not self.check_1_1().is_valid:
-            return LintResult.gen_simple_error_result(
-                "ファイルが読み込めなかったため、チェックできませんでした。", is_valid=None)
-        return func(self, *args, **kwargs)
-
-    return wrapper
 
 
 class CSVLinter:
